% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fit}
\alias{fit}
\title{Fit log linear models with step and L0 penalty}
\usage{
fit(object, lambdarange = c(1e-05, 0.01), B = 50)
}
\arguments{
\item{object}{object made with \code{\link{simdat}}.}

\item{lambdarange}{range of the lambda values for the L0 penalties.}

\item{B}{number of lambda values in \code{lambdarange}.}
}
\value{
A list with elements
\itemize{
\item \code{out} data frame with per method the number of non-zero parameter estimates \code{df},
the number of effective degrees of freedom \code{def}, the BIC \code{bic}, the number of
zero beta parameters \code{par0} in the starting model, \code{est0} the number of which have
a zero estimate, \code{lambda} the value yielding the best BIC, and \code{nrlambda} its number in
the lambda sequence.
\item \code{bhats} data frame with the sampled and estimated beta parameters.
}
}
\description{
Fit log linear models with step and L0 penalty
}
\details{
The starting model is a hierachical model excluding the k-factor interactions for which
one or more of the constituting (k - 1)-interactions have a non-zero sufficient statistic.

The models with an L0 penalty include \code{L0a} and \code{Lob}. For the former the BIC is computed as
\eqn{-2*\ell(\beta)+edf*\log(n)}, where \code{edf} are the effective degrees of freedom of
a ridge penalty, while for the latter the BIC is computed as \eqn{-2*\ell(\beta)+df*\log(n)},
where \code{df} are the number parameters with an absolute value greater than 1e-7.
}
\examples{
fit(simdat())

}
\references{
Frommlet, F. and Nuel, G. (2016). An adaptive ridge regression procedure for
\emph{L0} regularization. \emph{PLOS ONE}, \strong{11} (2), \url{doi:10.1371/journal.pone.0148620}.
}
