% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{repfit}
\alias{repfit}
\title{Conduct a simulation}
\usage{
repfit(
  reps = 1000,
  cores = NULL,
  nvars,
  levels,
  n,
  betas,
  lambdarange = c(1e-04, 0.01),
  B = 50
)
}
\arguments{
\item{reps}{number of parallel executed repetitions}

\item{cores}{number of cores. Default is maximum available minus 2.}

\item{nvars, levels, n, betas}{see \code{\link{simdat}}.}

\item{lambdarange, B}{see \code{\link{fit}}.}
}
\value{
The data frame \code{out} in the list returned by \code{\link{fit}} for
the \code{reps} replication.
}
\description{
Run the function \code{\link{fit}} in parallel
}
\examples{

# cores is set to 2 to pass the R package check

repfit(reps = 5, cores = 2, nvars = 4, levels = 2, n = 500, betas = -1:1)

}
