% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simdat}
\alias{simdat}
\title{Simulate a multiway contingency table}
\usage{
simdat(
  nvars = 3,
  levels = 3,
  n = 1000,
  betas = seq(-1, 1, by = 0.5),
  seed = NULL
)
}
\arguments{
\item{nvars}{number of variables.}

\item{levels}{the number of levels of each of the variables.}

\item{n}{population size.}

\item{betas}{vector with values from which to draw the \code{beta} parameters.}

\item{seed}{seed for reproducibility.}
}
\value{
A lists with:
\itemize{
\item \code{data} data frame with the observed data.
\item \code{beta} vector with the true beta parameters.
\item \code{fmax} formula of the maximal model.
}
}
\description{
Generates a contingency table with the frequencies generated
by taking a multinomial random sample from a randomly
generated parameter vector \code{beta}.
}
\examples{
simdat(seed = 1)

}
