% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fit}
\alias{fit}
\title{Model selection with step and L0 penalty based on the BIC}
\usage{
fit(object, lambdarange = c(1e-04, 0.01), B = 50)
}
\arguments{
\item{object}{object made with \code{\link{simdat}}.}

\item{lambdarange}{range of the lambda values for the L0 penalties.}

\item{B}{number of lambda values in \code{lambdarange}.}
}
\value{
A list with elements
\itemize{
\item \code{out} data frame containing:
\itemize{
\item \code{method} model selection method
\item \code{ic} information criterion (AIC vs BIC)
\item \code{seed} seed for generating the data
\item \code{var.lev} number of variables and levels of the variables.
\item \code{brange} range of beta parameters
\item \code{lrange} sum of the argument \code{lambdarange}
\item \code{n} sample size
\item \code{nbeta} number of parameters in starting model
\item \code{edf} effective degrees of freedom
\item \code{logl} maximized log-likelihood
\item \code{dev} deviance
\item \code{gof} AIC/BIC value, depending on the argument \code{ic}
\item \code{par0} number of beta parameters with value 0
\item \code{to_0} number of parameters in \code{par0} estimated as 0
\item \code{budget} sum of absolute parameter estimates, excluding the intercept
\item \code{lambdanr} number of \code{lambda} in the sequence of the tested lambda values
}
\item \code{bhats} data frame with the sampled and estimated beta parameters.
}
}
\description{
Model selection with step and L0 penalty based on the BIC
}
\details{
The starting model is a hierarchical model including the k-factor interaction parameters
for which the sufficient statistic is zero, provided that all corresponding lower-order
interaction parameters have positive sufficient statistics.

The effective degrees of freedom for the L0 method are calculated in the same way as
for the L2 ridge penalty. For the step method these are simply the number of non-zero
parameters in the model.
}
\examples{
fit(simdat())

}
\references{
Frommlet, F. and Nuel, G. (2016). An adaptive ridge regression procedure for
\emph{L0} regularization. \emph{PLOS ONE}, \strong{11} (2), \url{doi:10.1371/journal.pone.0148620}.
}
