% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{simdat}
\alias{simdat}
\title{Generate a multi-way contingency table}
\usage{
simdat(nvars = 3, levels = 3, n = 1000, betas = -1:1, seed = sample(1e+05, 1))
}
\arguments{
\item{nvars}{number of variables.}

\item{levels}{the number of levels of each of the variables.}

\item{n}{population size.}

\item{betas}{vector with values from which to draw the \code{beta} parameters.}

\item{seed}{seed for reproducibility.}
}
\value{
A lists with:
\itemize{
\item \code{data} data frame with the observed data.
\item \code{beta} vector with the true beta parameters.
\item \code{fmax} formula of the maximal model.
\item \code{seed} the seed used for the data simulation.
}
}
\description{
Generates a contingency table with the frequencies generated
by taking a multinomial random sample from a randomly
generated parameter vector \code{beta}.
}
\examples{
simdat(seed = 1)

}
